+++ lib/services/app_settings_service.dart
@@ -5,7 +5,6 @@ class AppSettingsService extends ChangeNotifier {
   static const _addKompressoPrefixKey = 'add_kompresso_prefix';
   static const _showOverheatWarningKey = 'show_overheat_warning';
   static const _saveVideosToAlbumKey = 'save_videos_to_album';
-  static const _deleteOriginalsAfterSavingKey = 'delete_originals_after_saving';
   static const _preventScreenSleepKey = 'prevent_screen_sleep';
   static const _languageCodeKey = 'language_code';
   static const _darkThemeKey = 'dark_theme';
@@ -20,7 +19,6 @@ class AppSettingsService extends ChangeNotifier {
   bool addKompressoPrefix = true;
   bool showOverheatWarning = true;
   bool saveVideosToAlbum = false;
-  bool deleteOriginalsAfterSaving = false;
   bool preventScreenSleep = true;
   String? languageCode;
   bool? darkTheme;
@@ -40,9 +38,6 @@ class AppSettingsService extends ChangeNotifier {
         _preferences.getBool(_showOverheatWarningKey) ?? showOverheatWarning;
     saveVideosToAlbum =
         _preferences.getBool(_saveVideosToAlbumKey) ?? saveVideosToAlbum;
-    deleteOriginalsAfterSaving =
-        _preferences.getBool(_deleteOriginalsAfterSavingKey) ??
-        deleteOriginalsAfterSaving;
     preventScreenSleep =
         _preferences.getBool(_preventScreenSleepKey) ?? preventScreenSleep;
     languageCode = _preferences.getString(_languageCodeKey);
@@ -57,12 +52,6 @@ class AppSettingsService extends ChangeNotifier {
 
   Future<void> setSaveVideosToAlbum(bool value) =>
       _setBool(_saveVideosToAlbumKey, value, (v) => saveVideosToAlbum = v);
-
-  Future<void> setDeleteOriginalsAfterSaving(bool value) => _setBool(
-    _deleteOriginalsAfterSavingKey,
-    value,
-    (v) => deleteOriginalsAfterSaving = v,
-  );
 
   Future<void> setPreventScreenSleep(bool value) =>
       _setBool(_preventScreenSleepKey, value, (v) => preventScreenSleep = v);